consolidated-frontend/requirements.txt
fastapi>=0.115.14
uvicorn[standard]>=0.25.0
chainlit>=2.0.1
azure-identity>=1.19.0
azure-ai-projects>=1.0.0b4
azure-communication-callautomation>=1.3.0
azure-eventgrid>=4.18.1
azure-cognitiveservices-speech>=1.40.0
aiohttp>=3.9.1
python-dotenv>=1.0.1
pydantic>=2.5.2
jinja2>=3.1.2
starlette>=0.41.3
requests>=2.32.3
pydub>=0.25.1
